requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Migrate==4.0.4
Werkzeug==2.3.6
tensorflow==2.12.0
numpy==1.24.2
Pillow==9.5.0
opencv-python==4.8.0.74
matplotlib==3.7.1
